requirements.txt
nextcord==2.6.0
python-dotenv==1.0.1
PyYAML==6.0.1
Requests==2.31.0
